[refactor]App名とTheme名を変更
