+++ KARTAR_Re_Make/app/src/main/java/com/example/kartar/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.example.kartar_re_make.ui.theme
+package com.example.kartar.ui.theme
 
 import androidx.compose.ui.graphics.Color
