+++ KARTAR_Re_Make/app/src/main/java/com/example/kartar/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.example.kartar_re_make.ui.theme
+package com.example.kartar.ui.theme
 
 import android.app.Activity
 import android.os.Build
@@ -38,7 +38,7 @@ private val LightColorScheme = lightColorScheme(
 )
 
 @Composable
-fun KARTAR_Re_MakeTheme(
+fun KARTARTheme(
     darkTheme: Boolean = isSystemInDarkTheme(),
     // Dynamic color is available on Android 12+
     dynamicColor: Boolean = true,
