+++ KARTAR_Re_Make/app/src/main/java/com/example/kartar/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.example.kartar_re_make.ui.theme
+package com.example.kartar.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
